Remove copy assembly code.
