+++ src/hotspot/cpu/aarch64/copy_aarch64.hpp
@@ -26,8 +26,6 @@
 #ifndef CPU_AARCH64_COPY_AARCH64_HPP
 #define CPU_AARCH64_COPY_AARCH64_HPP
 
-#include OS_CPU_HEADER(copy)
-
 static void pd_fill_to_words(HeapWord* tohw, size_t count, juint value) {
   julong* to = (julong*) tohw;
   julong  v  = ((julong) value << 32) | value;
@@ -52,7 +50,48 @@ static void pd_zero_to_bytes(void* to, size_t count) {
   (void)memset(to, 0, count);
 }
 
-#ifndef _WINDOWS
+#ifdef _WINDOWS
+
+static void pd_conjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
+  (void)memmove(to, from, count * HeapWordSize);
+}
+
+static void pd_disjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
+  switch (count) {
+  case 8:  to[7] = from[7];
+  case 7:  to[6] = from[6];
+  case 6:  to[5] = from[5];
+  case 5:  to[4] = from[4];
+  case 4:  to[3] = from[3];
+  case 3:  to[2] = from[2];
+  case 2:  to[1] = from[1];
+  case 1:  to[0] = from[0];
+  case 0:  break;
+  default:
+    (void)memcpy(to, from, count * HeapWordSize);
+    break;
+  }
+}
+
+static void pd_disjoint_words_atomic(const HeapWord* from, HeapWord* to, size_t count) {
+  switch (count) {
+  case 8:  to[7] = from[7];
+  case 7:  to[6] = from[6];
+  case 6:  to[5] = from[5];
+  case 5:  to[4] = from[4];
+  case 4:  to[3] = from[3];
+  case 3:  to[2] = from[2];
+  case 2:  to[1] = from[1];
+  case 1:  to[0] = from[0];
+  case 0:  break;
+  default: while (count-- > 0) {
+             *to++ = *from++;
+           }
+           break;
+  }
+}
+
+#else
 
 #define COPY_SMALL(from, to, count)                                     \
 {                                                                       \
@@ -133,7 +172,7 @@ static void pd_conjoint_words(const HeapWord* from, HeapWord* to, size_t count) 
     COPY_SMALL(from, to, count);
     return;
   }
-  _Copy_conjoint_words(from, to, count);
+  (void)memmove(to, from, count * HeapWordSize);
 }
 
 static void pd_disjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
@@ -146,7 +185,7 @@ static void pd_disjoint_words(const HeapWord* from, HeapWord* to, size_t count) 
     COPY_SMALL(from, to, count);
     return;
   }
-  _Copy_disjoint_words(from, to, count);
+  memcpy(to, from, count * HeapWordSize);
 }
 
 static void pd_disjoint_words_atomic(const HeapWord* from, HeapWord* to, size_t count) {
@@ -155,15 +194,18 @@ static void pd_disjoint_words_atomic(const HeapWord* from, HeapWord* to, size_t 
     COPY_SMALL(from, to, count);
     return;
   }
-  _Copy_disjoint_words(from, to, count);
+  // ??? not sure if this one's right.
+  memcpy(to, from, count * HeapWordSize);
+}
+
+#endif // _WINDOWS
+
+static void pd_aligned_disjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
+  pd_disjoint_words(from, to, count);
 }
 
 static void pd_aligned_conjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
   pd_conjoint_words(from, to, count);
-}
-
-static void pd_aligned_disjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
-  pd_disjoint_words(from, to, count);
 }
 
 static void pd_conjoint_bytes(const void* from, void* to, size_t count) {
@@ -175,43 +217,75 @@ static void pd_conjoint_bytes_atomic(const void* from, void* to, size_t count) {
 }
 
 static void pd_conjoint_jshorts_atomic(const jshort* from, jshort* to, size_t count) {
-  _Copy_conjoint_jshorts_atomic(from, to, count);
+  if (from > to) {
+    while (count-- > 0) {
+      // Copy forwards
+      *to++ = *from++;
+    }
+  } else {
+    from += count - 1;
+    to   += count - 1;
+    while (count-- > 0) {
+      // Copy backwards
+      *to-- = *from--;
+    }
+  }
 }
 
 static void pd_conjoint_jints_atomic(const jint* from, jint* to, size_t count) {
-  _Copy_conjoint_jints_atomic(from, to, count);
+  if (from > to) {
+    while (count-- > 0) {
+      // Copy forwards
+      *to++ = *from++;
+    }
+  } else {
+    from += count - 1;
+    to   += count - 1;
+    while (count-- > 0) {
+      // Copy backwards
+      *to-- = *from--;
+    }
+  }
 }
 
 static void pd_conjoint_jlongs_atomic(const jlong* from, jlong* to, size_t count) {
-  _Copy_conjoint_jlongs_atomic(from, to, count);
+  pd_conjoint_oops_atomic((const oop*)from, (oop*)to, count);
 }
 
 static void pd_conjoint_oops_atomic(const oop* from, oop* to, size_t count) {
-  assert(BytesPerLong == BytesPerOop, "jlongs and oops must be the same size");
-  _Copy_conjoint_jlongs_atomic((const jlong*)from, (jlong*)to, count);
+ if (from > to) {
+    while (count-- > 0) {
+      // Copy forwards
+      *to++ = *from++;
+    }
+  } else {
+    from += count - 1;
+    to   += count - 1;
+    while (count-- > 0) {
+      // Copy backwards
+      *to-- = *from--;
+    }
+  }
 }
 
 static void pd_arrayof_conjoint_bytes(const HeapWord* from, HeapWord* to, size_t count) {
-  _Copy_arrayof_conjoint_bytes(from, to, count);
+  pd_conjoint_bytes_atomic(from, to, count);
 }
 
 static void pd_arrayof_conjoint_jshorts(const HeapWord* from, HeapWord* to, size_t count) {
-  _Copy_arrayof_conjoint_jshorts(from, to, count);
+  pd_conjoint_jshorts_atomic((const jshort*)from, (jshort*)to, count);
 }
 
 static void pd_arrayof_conjoint_jints(const HeapWord* from, HeapWord* to, size_t count) {
-   _Copy_arrayof_conjoint_jints(from, to, count);
+  pd_conjoint_jints_atomic((const jint*)from, (jint*)to, count);
 }
 
 static void pd_arrayof_conjoint_jlongs(const HeapWord* from, HeapWord* to, size_t count) {
-  _Copy_arrayof_conjoint_jlongs(from, to, count);
+  pd_conjoint_jlongs_atomic((const jlong*)from, (jlong*)to, count);
 }
 
 static void pd_arrayof_conjoint_oops(const HeapWord* from, HeapWord* to, size_t count) {
-  assert(!UseCompressedOops, "foo!");
-  assert(BytesPerLong == BytesPerOop, "jlongs and oops must be the same size");
-  _Copy_arrayof_conjoint_jlongs(from, to, count);
+ pd_conjoint_oops_atomic((const oop*)from, (oop*)to, count);
 }
-#endif // _WINDOWS
 
 #endif // CPU_AARCH64_COPY_AARCH64_HPP
